requirements.txt
# LLM Red Teamer — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core HTTP Client ──────────────────────────────────────────────────────────
httpx>=0.27.0              # Async-capable HTTP client with retry support

# ── CLI / Terminal UI ─────────────────────────────────────────────────────────
rich>=13.7.0               # Terminal formatting, tables, progress bars

# ── Payload Library ───────────────────────────────────────────────────────────
pyyaml>=6.0.1              # YAML payload library parsing

# ── Dashboard Backend ─────────────────────────────────────────────────────────
fastapi>=0.111.0           # Web framework for the dashboard API
uvicorn[standard]>=0.29.0  # ASGI server for FastAPI
python-multipart>=0.0.9    # Required for FastAPI file upload

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=8.2.0              # Test runner
pytest-mock>=3.14.0        # Mocking utilities
